+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#904](https://github.com/NibiruChain/nibiru/pull/904) - refactor: change Pool name to VPool in vpool module
 * [#894](https://github.com/NibiruChain/nibiru/pull/894) - add the collections package!
 * [#897](https://github.com/NibiruChain/nibiru/pull/897) - x/pricefeed - use collections.
+* [#933](https://github.com/NibiruChain/nibiru/pull/933) - refactor(perp): remove whitelist and simplify state keys
 
 ### API Breaking